allow bundling of @wdio/repl refs #5139
